backend/ai/requirements.txt
pymongo[srv]
python-dotenv
requests
numpy